Remove unnecessary NO_ASSERTS macro use
Remove unnecessary NO_ASSERTS macro use, as utils.h already checks the
NO_ASSERTS macro and changes the assert macros to empty expressions.
Therefore it is not necessary to fence the usage of assert, as the dead
code elimiation of the compiler should optimize out stuff like unused
conditional branches.

Furthermore remove the cassert include, as we use a custom assert macro